Update UI of etching part in rune-crafter

+++ src/pages/rune-crafter/ui/form/EtcherTab/validation.ts
@@ -11,11 +11,11 @@ const schema = z.object({
 		.max(28, 'Max 28 chars'),
 	runeSymbol: z.string().length(1, 'Only one symbol'),
 	destAddress: z.string().min(3),
-	divisibility: z.coerce.number().optional(),
-	premine: z.coerce.number().optional(),
+	divisibility: z.coerce.number().int().optional(),
+	premine: z.coerce.number().int().optional(),
 	mintType: z.enum(['open', 'closed']),
-	mintAmount: z.coerce.number().optional(),
-	mintCap: z.coerce.number().optional(),
+	mintAmount: z.coerce.number().int().optional(),
+	mintCap: z.coerce.number().int().optional(),
 })
 
 export function useFormValidation() {
